+++ AABB_tree/doc/AABB_tree/Concepts/AABBRayIntersectionGeomTraits_3.h
@@ -2,7 +2,7 @@
 \ingroup PkgAABBTreeConcepts
 \cgalConcept
 
-The concept `AABBRayIntersectionGeomTraits` is a refinement of the
+The concept `AABBRayIntersectionGeomTraits_3` is a refinement of the
 concept `AABBGeomTraits`. In addition to the types required by
 `AABBGeomTraits` it also requires types and functors necessary to
 define the Intersection_distance functor.
@@ -13,12 +13,12 @@ define the Intersection_distance functor.
 \cgalHasModelsBare{All models of the concept `Kernel`}
 \cgalHasModelsEnd
 
-\sa `CGAL::AABB_traits<AABBGeomTraits,AABBPrimitive>`
+\sa `CGAL::AABB_traits_3<AABBGeomTraits,AABBPrimitive>`
 \sa `CGAL::AABB_tree<AABBTraits>`
 \sa `AABBPrimitive`
 
 */
-class AABBRayIntersectionGeomTraits {
+class AABBRayIntersectionGeomTraits_3 {
 public:
   /*!
     Type of a 3D ray.