+++ AABB_tree/test/AABB_tree/aabb_distance_edge_test.cpp
@@ -21,7 +21,7 @@
 #include <CGAL/Simple_cartesian.h>
 
 #include <CGAL/AABB_tree.h>
-#include <CGAL/AABB_traits.h>
+#include <CGAL/AABB_traits_3.h>
 #include <CGAL/Polyhedron_3.h>