+++ AABB_tree/test/AABB_tree/aabb_naive_vs_tree_distance_triangle_test.cpp
@@ -27,7 +27,7 @@
 #include <CGAL/Polyhedron_3.h>
 
 #include <CGAL/AABB_tree.h>
-#include <CGAL/AABB_traits.h>
+#include <CGAL/AABB_traits_3.h>
 
 #include "AABB_test_util.h"