+++ Minkowski_sum_2/include/CGAL/Minkowski_sum_2/AABB_collision_detector_2.h
@@ -37,7 +37,7 @@ public:
   typedef typename Polygon_2::Edge_const_iterator       Edge_iterator;
   typedef AABB_segment_2_primitive<Kernel, Edge_iterator, Polygon_with_holes_2>
                                                         Tree_segment_2;
-  typedef AABB_traits_2<Kernel, Tree_segment_2>         Tree_traits;
+  typedef Minkowski_sum::AABB_traits_2<Kernel, Tree_segment_2>         Tree_traits;
   typedef AABB_tree_with_join<Tree_traits>              Tree_2;
 
 public: